-Added new objects to handle parsing and store an Rdio collection. -Added new UITableView delegate that utilizes the Rdio.framework to download, parse, and browse through an Rdio user's collection. -Added a new method to SonosController.h to send an Rdio song for the Sonos Controller to play.
Please note this is mainly proof concept of browsing a users Rdio library and sending a selected Rdio track to a Sonos controller. Downloading and parsing Sonos collection data can take a long time. (For my 3200 track library it takes upwards of 30 seconds to download and parse the entire collection).