Adding NBAnimation which is a wrapper for CAKeyframeAnimation. This is a forked version of SKBounceAnimation by @khanlou.
Moving PLDirectionGestureRecognizer into NBKit and renaming it to NBDirectionGestureRecognizer. At some point I'll wrap this stuff up nicer so it can be reused in other projects.

Cleaned up some stuff in the NowPlaying view controller so the UI scales better when in landscape or eventually iPad. Still have some work to do here.