Trying out a new speaker interface. It's not completely realized here but the basic idea is you have control of individual speakers on the left and on the right you can "pair" speakers with whatever speaker was first introduced to that space. I've been using it for a couple weeks now and it's pretty fun.
I removed the "Playing" button from this view and it sucks because now you have to tap a speaker then "Playing" to see what's currently playing and gain access to volume. In practice, these controls are too important to be that many taps away so I plan to make this more accessible in the speaker view soon.